Split classes and Fix #264
