+++ src/CodeFormatter/Operation.cs
@@ -1,0 +1,9 @@
+﻿namespace CodeFormatter
+{
+    public enum Operation
+    {
+        Format = 0,
+        ListRules = 1,
+        ShowHelp = 2
+    }
+}
